stream Manager: Added code for influx DB subscription.
the stream Manager is started by a command named "da"
This starts a UDP server and creates subscription in influxDB
Data recieved by influx need to be forwarded to data bus.

KNOWN BUG: The logging infra is throuwing a parse error which will
be resolved in next release and it is not that critical.

TODO:
1. Parse the influx point and read from respective storage
   Forward the same to data-bus.
2. Make logging infra more usable and stable.
3. Use golang unit test framework, today it uses a raw executable.
4. Use authentication cred for using influx and redis.
5. Export the data to file system or any other external medium.

Change-Id: I162ac5808919c6cdb85b3ae69135134266657c0f
Signed-off-by: Das, Lalatendu <lalatendu.das@intel.com>