DataAnalytics: Enable YuMei classification with Kapacitor
This patch enables the basic classification with the Yumei classifier
algorithm with Kapacitor.
The classification algo needs python 3.6 and the kapacitor currently
supports only python 2.7. For this currently a linux socket is opened
for communication between the python 2.7 udf and python 3.6 classifier.

This will be changed in future with a modified version Kapacitor which
can work with udf in python 3.x.

Change-Id: Ic526b4512d0350bdf68447317f16cdadc4c95aad
Signed-off-by: Arun Kumar K <arun1.k.kumar@intel.com>
Signed-off-by: Subramaniam, Sathiyamoorthy <sathiyamoorthy.subramaniam@intel.com>