DataIngestionLib, VideoIngestion: Some optimizations (#161)
* DataIngestionLib, VideoIngestion: Some optimizations

The VideoIngesiton was creating a new object of DIL every time a trigger
is received. This was causing lot of overhead of calling GetConfigInt,
Get ImageStoreClient certs, call into image store etc on every frame in
case of PCB demo.
This patch optimizes the DIL to have an independent object called data
point which can be intialized independent of DataIngestionLib.
This is also tested for multi threading by more number of trigger_threads.

This patch fixes the system hang issue in camera mode. Also removes the
sleep in file mode ingestion.

Signed-off-by: Arun Kumar K <arun1.k.kumar@intel.com>