[HACK] DataBusAbstraction: Fix for making long run work
Overnight run of ETA with pcb demo fails after many hours with error:
UA_NodeMap_releaseNode: Assertion 'entry->refCount > 0' failed.

This can be easily reproduced in few seconds by running the go publish
in a continous loop and running python subscribe in a loop in another
terminal.

Issue seems to be due to a bug in open62541 causing it to fail if node
updates happen in quick succession.

This patch adds a sleep of 10ms before every write and this is making
the continous publish subscribe loop run long time without failures.
This needs to be properly analyzed and fixed.

This patch also removes some prints which clogs the publish - subscribe
cycle.

Signed-off-by: Arun Kumar K <arun1.k.kumar@intel.com>