DataAnalytics, docker_setup: Some fixes for long run
For PCB demo, reducing the retention time for the images in redis db.
Due to this there can be issues sometimes of images being referred in
DataAnalytics post retention time. So this patch also fixes stoppage of
data analytics if one image frame is not receivied, making it continue
to the next frame.

Also increasing the number of trigger threads to 12 to account for
certain systems where still memory increase is observed.

Signed-off-by: Arun Kumar K <arun1.k.kumar@intel.com>