Docker_Setup: Fixed the HOST_TIME_ZONE args not consumed warning (#203)
* Docker_Setup: Fixed the HOST_TIME_ZONE args not consumed warning

HOST_TIME_ZONE is passed as an argument only to ia-pybase and
ia-gobase as other modules use those as a base image. The local
machine time_zone setting logic is executed only in ia-pybase and
ia-gobase. For any other new container which doesn't use ia-gobase
and ia-pybase the time_zone setting logic must be applied and
HOST_TIME_ZONE must be passed as a build argument.

Signed-off-by: sandeshk <sandesh.kumar.s.s@intel.com>