StreamManager: Implement wait for publishing points (#224)
Until the refCount issue is fixed, we have to slow down the publish. The
max speed of ETA publish will be limited and if the classifier works
faster than that, the points get discarded with an Info message.

Made a channel between http handler func and publish routine. When the
channel is full, the points get discarded. Introduced a 500ms delay
after doing every publish.

Signed-off-by: Arun Kumar K <arun1.k.kumar@intel.com>