VideoIngestion: Hitachi RTSP H.264 issue fix
Potential fix to RTSP issue found by Hitachi when attempting to use
H.264 decoder to ingest an RTSP stream. The issue seemed to be user
access to the GPU, which came from the ieiuser not actually exist in the
container itself. So the code was ran as no user and so the user in the
container didn't have access to the GPU.

Signed-by: Kevin Midkiff <kevin.midkiff@intel.com>