EISMessageBus: Disconnect WARN and Authentication Exception Add
Changes the disconnection functionality in the EIS Message Bus to only log the disconnect as a warning and keep the socket alive with no exceptions raised. Additionally, this adds an exception which is raised in all of the receive methods across the C, Python, and Go libraries for if the authentication fails for the socket. There are additionally a couple of other small fixes to some random issues I found while testing.

Impact:
Since the receive method now raises an error for authentication failures for the receive methods, the services using the EIS Message Bus must respond to this error, either by exiting or whatever mitigation strategy we have for this.

Signed-off-by: Kevin Midkiff <kevin.midkiff@intel.com>