EISMessageBus: Send Message SEGFAULT Issue
Bug: If an error occurred when sending a message (publishing or issuing
a response) a SEGFAULT would occur because a double free was happening
in the error handling code of the underlying send_message() helper
function.

Fix: This commit fixes the bug above by changing how the underlying error
handling happens in the send_message() function in the ZMQ protocol
wrapper. Additionally, this merge correctly propagates errors which
occur in this method up to the user.

Signed-off-by: Midkiff, Kevin <kevin.midkiff@intel.com>