EIS: Removed `read_only: true` config from VI service
Removed `read_only: true` config key from VideoIngestion
docker-compose service as it was found that without this key,
inferencing on `MYRIAD` device will not work.

Even with this fix, it was found that gvadetect device=MYRIAD
still doesn't work but this fix should enable running of
classifier UDF with inference happening on `MYRIAD` device.
However, gvatect with devices: CPU, GPU and HDDL seems to work
as expected without any issues

Signed-off-by: Vinod K B <vinod.k.b@intel.com>