Discover_History: Changing the receiving syntax of service.recv() function call.
Observing a weird behavior, sometimes response = service.recv() is working fine and sometimes it
is throwing error saying "TypeError: tuple indices must be integers or slices, not str"

Same goes with the response, _ = service.recv(). It is working fine sometimes and sometimes it throws
error saying "ValueError: not enough values to unpack (expected 2, got 1)"

The root cause of the issue could be the eis version in the .env file.
As discover history is dependent on ia_common:{EIS_VERSION} with 2.1 in the .env,
we are getting the data with response = service.recv().

With version set to 2.2 in the .env,
we are getting the data with response, _ = service.recv().

Changes:
1. Addressed pycodestyle warnings
2. Update the version to 2.2 in the .env file

Signed-off-by: Shubham Kar <shubhamx.kar@intel.com>