docker_setup: updated the volume mount of Altera.icd for VI and VA
Instead of volume mounting the entire `/etc/OpenCL/vendors` folder
inside container which was mounting an empty folder overriding
the container filesystem which had intel.icd there, we are just
volume mounting the required `Altera.icd` file which gets installed
on fpga configuration on host in VI and VA services. On host system
where fpga is not configured, the `Altera.icd` will be an empty
directory whereas if fga is configured, `Altera.icd` would
be a file.

Signed-off-by: Vinod K B <vinod.k.b@intel.com>