eis-core:Changes done to add provisoning by K8
for EIS components and enable video use case and

sample python publisher-subscriber.

1.K8 secrets have been added
for apps and etcd in common gen_certs.py file

2.Follow Readme_k8 for understanding basic steps.
It is only for help in review but is not final.
Separate story exists for README.

3.Added K8s provioning mode in provisioning script.

4.Created one deployment script for both dev and prod
mode provisoning.

5. Added sample app pub-sub Yml files & video use case.
WebVisualizer is added in next story.

6. k8_eis/convert_orig_final.sh is not final yet.
Changes will be added more.

7. Added k8s folder to contain all k8s related yml and script files.

Signed-off-by: Vatika <vatika.gupta@intel.com>