build: EIS to build UDF service placed outside of IEdgeinsights.
This patch enables user to define the service as a full path name in yml file
and eis_builder.py to pick the service to generate cosolidated
docker-compose and config.json file. This feature is needed if Algo
develepor wanted to build the custom UDF completely outside of EIS directory
structure.

Signed-off-by: Lalatendu Das <lalatendu.das@intel.com>