EISUtils: Add ThreadExecutor utility
Add ThreadExecutor utility to be used by the UDFManager to eliminate the need for the extra job queue associated with using the ThreadPool utility. This also contains a slight fix to the thread safe queue, and there being an instance where it was not entirely thread-safe.

These improvements are associated with the JIRA ticket [IAPOC9-3115](https://jira.devtools.intel.com/browse/IAPOC9-3115).

Signed-off-by: Kevin Midkiff <kevin.midkiff@intel.com>