eis-core: Added fast base64 library to utils
The fast base64 encoder/decoder library
(https://github.com/aklomp/base64) has been
added to utils. This is needed to enable
depth frames in the video pipeline as currently
the depth frames will be encoded using this
library and added to frame metadata. The udf
will then have to use this library to decode
the frame.

Signed-off-by: sandeshk1 <sandesh.kumar.s@intel.com>