common: Installing grpc and protobuf packages using .deb file
Also, removal of cloning grpc and its installation via install.sh

Reason behind this fix is that, our EII was installing grpc and protobuf
libraries by cloning the grpc repo which internally clones from multiple
other repos. This was a problem in PRC network as many of the links were
not available and was not able to get the mirror links.

Signed-off-by: sagartp <sagar.tp@intel.com>