build/provision: Fix system IP in  provisioning script
When HOST_IP is not provided in .env provisioning
script tries to get the System IP using a command.
This command was getting the wrong IP address. This
issue was noticed when working with GigE camera when
when the HOST IP needed to be added to ETCD HOST and
no_proxy. VI was not able to connect to etcd as provisioing
was executed with a diff System IP.

To fix this issue the command which fetched the system IP
was modified by updating the grep command to fetch the IP
from the correct network interface.

Signed-off-by: sandeshk1 <sandesh.kumar.s@intel.com>