build/builder.py: Fixed an issue where ia_video_common always builds 		  even when time-series usecase is built
The ia_video_common service builds for even time_series use case, which not needed and a waste of time.
In future the same issue can happen with a new base service, which should be avoided as
these builds takes considerable time when built with --no-cache.

The solution here, checks the services and their (nested) dependencies and choose the only
base services which are required.

changes made:

	1. Modified builder.py to include only the required services in docker-compose-build.yml
	   by checking the services dependencies

Signed-off-by: Safeer Vengasseri <safeerx.vengasseri@intel.com>