build: Fix HDDL issue on ubuntu 20
In ubuntu 20 systems HDDL initialization was failing
as eiiuser was being inside the video-usecase docker
services but the hddldaemon files were generated with
the user of the host system.

During runtime volume mount of hddl files under /var/tmp
there was a conflict in user name of the hddl files
and the user used inside docker service.

Hence a readme step is added to change ownership of
hddl files on host system to eiiuser. Post volume mount
the ownership of hddl files inside docker container is
set to eiiuser.

Signed-off-by: sandeshk1 <sandesh.kumar.s@intel.com>