build: Updated .env to source HTTP_METHOD to switch in RDE
This patch contains the changes in .env and values.yml of
helm charts to get the swtich parameter of HTTP_METHOD.
We are enabling only GET call as default.
POST call needs a pre-requisite step of external HTTP server running.
With this parameter RDE will start the Server to server only
GET or GET with POSTING the data to extrernal Server

Signed-off-by: Vivek R Soundararajan <vivek.rs@intel.com>